completed compatibility with ledController.rs
